added booster token verifier class
