+++ packages/framework-core/test/booster-auth.test.ts
@@ -8,8 +8,6 @@ import { Logger, BoosterConfig } from '@boostercloud/framework-types'
 import { RoleAccess, UserEnvelope } from '@boostercloud/framework-types'
 import { BoosterAuth } from '../src/booster-auth'
 import { ProviderLibrary } from '@boostercloud/framework-types'
-import createJWKSMock from 'mock-jwks'
-import { internet, phone, random } from 'faker'
 
 const logger: Logger = {
   debug() {},
@@ -251,66 +249,5 @@ describe('the "isUserAuthorized" method', () => {
     }
 
     expect(BoosterAuth.isUserAuthorized(authorizedRoles, userEnvelope)).to.eq(true)
-  })
-})
-
-describe('the "verifyToken" method', () => {
-  const auth0VerifierUri = 'https://myauth0app.auth0.com/'
-  const issuer = 'auth0'
-  const jwks = createJWKSMock(auth0VerifierUri)
-  const email = internet.email()
-  const phoneNumber = phone.phoneNumber()
-  const userId = random.uuid()
-  const config = new BoosterConfig('test')
-  config.tokenVerifier = {
-    issuer,
-    jwksUri: auth0VerifierUri + '.well-known/jwks.json',
-  }
-
-  beforeEach(() => {
-    jwks.start()
-  })
-
-  afterEach(async () => {
-    await jwks.stop()
-  })
-
-  it('decode and verify an auth token with the custom roles', async () => {
-    const token = jwks.token({
-      sub: userId,
-      iss: issuer,
-      'custom:role': 'User',
-      email: email,
-      phone_number: phoneNumber,
-    })
-
-    const expectedUser: UserEnvelope = {
-      id: userId,
-      username: email,
-      role: 'User',
-    }
-
-    expect(await BoosterAuth.verifyToken(config, token)).to.deep.equals(expectedUser)
-  })
-
-  it('fails if a different issuer emitted the token', async () => {
-    const token = jwks.token({
-      iss: 'firebase',
-    })
-
-    await expect(BoosterAuth.verifyToken(config, token)).to.eventually.be.rejected
-  })
-
-  it('fails if a token has expired', async () => {
-    const token = jwks.token({
-      sub: userId,
-      iss: issuer,
-      'custom:role': 'User',
-      email: email,
-      phone_number: phoneNumber,
-      exp: 0,
-    })
-
-    await expect(BoosterAuth.verifyToken(config, token)).to.eventually.be.rejected
   })
 })
